Update to Unreal Engine 5.5
I can't figure out why but Installing the Visual Studio Integration Tool doesn't work if it's installed to the Engine, it has to be added to the project to work. The VS error is very generic, so need to spend more time later to troubleshoot so the plugin can be removed from this project and installed to the editor.